hw_2/requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==24.3.0
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
frozenlist==1.5.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kaleido==0.2.1
magic-filter==1.0.12
multidict==6.1.0
packaging==24.2
plotly==5.24.1
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
